어드민 추천 조회 score Decimal 일치 및 유저 존재 검사 최적화

+++ apps/recommendations/serializers.py
@@ -146,7 +146,7 @@ class RecommendationJobRunResponseSerializer(serializers.Serializer):  # type: i
 
 class AdminUserRecommendationItemSerializer(serializers.Serializer):  # type: ignore[type-arg]
     game_id = serializers.IntegerField(source="igdb_game_id")
-    score = serializers.FloatField()
+    score = serializers.DecimalField(max_digits=5, decimal_places=4)
 
 
 class AdminUserRecommendationListResponseSerializer(serializers.Serializer):  # type: ignore[type-arg]
